fix name lookup issues for gcc

+++ dctl/walk/Root.hpp
@@ -117,7 +117,7 @@ private:
                 std::cout << " time ";
                 std::cout << std::setw( 6) << timer.lap();
 
-                double const nps = (1000 * statistics_.nodes()) / static_cast<double>(timer.lap());
+                double const nps = static_cast<double>(1000 * statistics_.nodes()) / static_cast<double>(timer.lap());
                 std::cout << " nps ";
                 std::cout << std::dec << std::setiosflags(std::ios::fixed) << std::setprecision(0);
                 std::cout << std::setw( 7) << nps;
